+++ packages/hcl-sdk-web-ui/src/components/ui-kits/hcl-sdk-icon-flag/FI.tsx
@@ -1,0 +1,12 @@
+import { h } from '@stencil/core';
+
+export default () => (
+  <svg width="1em" height="1em" version="1.1" xmlns="http://www.w3.org/2000/svg" viewBox="37.554 0 341.3 341.3">
+    <rect y="0" fill="#FFFFFF" width="512" height="341.3" />
+    <polygon
+      fill="#2E52B2"
+      points="512,129.3 512,212 203.7,212 203.7,341.3 121,341.3 121,212 0,212 0,129.3 121,129.3 121,0 203.7,0
+	203.7,129.3 "
+    />
+  </svg>
+);