+++ packages/hcl-sdk-web-ui/src/components/ui-kits/hcl-sdk-icon-flag/IE.tsx
@@ -1,0 +1,9 @@
+import { h } from '@stencil/core';
+
+export default () => (
+  <svg width="1em" height="1em" xmlns="http://www.w3.org/2000/svg" viewBox="85.333 85.333 341.333 341.333">
+    <rect y="85.331" fill="#FFFFFF" width="512" height="341.326" />
+    <rect y="85.331" fill="#6DA544" width="170.663" height="341.337" />
+    <rect x="341.337" y="85.331" fill="#FF9811" width="170.663" height="341.337" />
+  </svg>
+);