+++ packages/hcl-sdk-web-ui/src/core/stores/FeatureStore.ts
@@ -1,3 +1,3 @@
 export const featureStore = {
-  isPatientRecommendationEnabled: false
+  isPatientRecommendationEnabled: true
 }