+++ packages/hcl-sdk-web-ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 0.3.0
+
+### Minor Changes
+
+- cc774b3: minor features and bug fixes:
+
+  - showing type in hco suggestion results
+  - format url for web address #977
+  - handle missing specialties in hcp individual #1165
+  - loading indicator for map #1285
+  - hide relaunch btn after change currentPosition #1286
+  - update suggestion modification link #978
+  - fix appointment link #1368
+  - add country flags #1330
+
 ## 0.2.0
 
 ### Minor Changes